Refactor YAML validation logic from magefile.go to a new file schemaValidator.go.
- Fixed failures to catch problems in editsteps and args
